feat(hooks): create usePersonalDataSave hook for managing personal data saving logic
