+++ frontend/web/src/features/guestdashboard/GuestSettings.js
@@ -1,5 +1,6 @@
-import React, { useState } from "react";
+import React from "react";
 import useSettingsData from "../../hooks/useSettingsData";
+import usePersonalDataSave from "../../hooks/usePersonalDataSave";
 import PersonalDataForm from "../hostdashboard/hostsettings/components/PersonalDataForm";
 import "../../styles/sass/pages/dashboard/settingsDashboard.css";
 import "../hostdashboard/hostsettings/styles/hostSettings.css";
@@ -9,51 +10,7 @@ const SHOW_AUTH_MFA = false;
 
 const GuestSettings = () => {
     const settingsData = useSettingsData();
-    const {
-        user,
-        tempUser,
-        selectedCountryCode,
-        stripPhone,
-        onSaveUserName,
-        onSaveUserEmail,
-        onSaveUserPhone,
-        onSaveUserDateOfBirth,
-        onSaveUserPlaceOfBirth,
-        onSaveUserNationality,
-    } = settingsData;
-
-    const [isSaving, setIsSaving] = useState(false);
-    const [saveSuccess, setSaveSuccess] = useState(false);
-
-    const saveAll = async () => {
-        const saves = [];
-
-        if ((tempUser.name || "").trim() !== (user.name || "").trim()) {
-            saves.push(onSaveUserName());
-        }
-        if ((tempUser.email || "").trim() !== (user.email || "").trim()) {
-            saves.push(onSaveUserEmail());
-        }
-        const fullPhone = `${selectedCountryCode}${stripPhone || ""}`.trim();
-        if (fullPhone !== (user.phone || "").trim()) {
-            saves.push(onSaveUserPhone());
-        }
-        if ((tempUser.dateOfBirth || "") !== (user.dateOfBirth || "")) {
-            saves.push(onSaveUserDateOfBirth());
-        }
-        if ((tempUser.placeOfBirth || "") !== (user.placeOfBirth || "")) {
-            saves.push(onSaveUserPlaceOfBirth());
-        }
-        if ((tempUser.nationality || "").trim() !== (user.nationality || "").trim()) {
-            saves.push(onSaveUserNationality());
-        }
-
-        setIsSaving(true);
-        await Promise.allSettled(saves);
-        setIsSaving(false);
-        setSaveSuccess(true);
-        setTimeout(() => setSaveSuccess(false), 2500);
-    };
+    const { saveAll, isSaving, saveSuccess } = usePersonalDataSave(settingsData);
 
     return (
         <PersonalDataForm
